Amending Constitution
Proposal 2:
Ammend section 5 of the constitution to read as follows:

5 § On Time Limits and Inactive Players

A player has to complete all mandated actions before the end of the next day (always 17:00 CET) unless stated otherwise.

If a player doesn’t fulfill an obligation compelled by the rules within the time granted for it, they are considered Inactive. Inactive players will be given the "Inactive" role, they will be removed from the turn order in #turns (thus their proposal turns will be skipped) and they do not have access to the #rule-voting channel. Players who comply with this rule are considered Active.

A player tagged as "Inactive" can make a request to become active again in the #nomic-discussion channel. Upon doing so, they are considered active again and are added in last place in #turns.

4 Yes
0 No