Removing section 7 adding section 9
Zephnik
Proposal 26: Delete the seventh section of the Constitution
Y 5  N 1 


- **7 § Transitional Arrangements**

The Constitution completely replaces Rule 1, Rule 2, Rule 3, Rule 5, Rule 8 and Rule 13.

Klink
Proposal 25 
9 § On Joining the Game 

When a discord user joins the server, they are not automatically added to the turns list, and are not considered to be playing the game. The user cannot gain points, vote, propose, or otherwise participate in the game outside of chatting.

To join the game, a user may use the !interested command and ping a @Moderator indicating intent to join. At this point, they are officially recognized as a player, and may be added to the turn list and otherwise allowed to participate in the game.