Adding and implementing rule 31
Create a rule titled "Rule Titles":

Each rule may have a title in addition to its number. If the proposal creating a rule specifies a title for the new rule, then that becomes the rule's title upon the proposal's adoption. A proposal to amend a rule may specify a new title for the rule; that title becomes the rule's title of the proposal takes effect. If a rule ever lacks a title, any Github collaborator may give it a title. 

If a proposal to create, amend or repeal a rule fails to include the title of the rule, 20 coins are transferred from the proposal creator to the Bank upon the proposal's adoption. If a proposal specifies a new title for a rule, the old title must be included as well to avoid the fine. 

For the sake of clarity, the constitution's title is  "The Constitution".
---
Upon this rule taking effect, give the following titles to the rules indicated. 

6: TripleVotes
7: Coins for passing proposals
9: Coins for voting against proposals
10: Debates 
14: Bribery 
16: Using balloons 
17: Economy 
18: Voluntary inactivity 
19: Vassals & Lords
20: Unconstitutional incidents 
21: Tax
22: Redistribution of wealth 
23: Proposal ID'S
24: #mask-archives
25: Errata
26: Loans 

The rule created by proposal 16 (if it passes): Pickpocketing
The rule created by proposal 20 (if it passes): Trades
The rule created by proposal 21 (if it passes): Wealth limits 

Once the above titles are applied, any collaborator may remove the "---", and everything following it, from this rule.