Proposals 36, 38, 39, 40
Proposal 36: modify rule 30: Wealth limits by removing the last paragraph Y3 N0
Proposal 38: If you are a @Current player, you may lose 20 coins. If you do, you may make 3 feudal actions today. Y4 N1
Proposal 39: Amend rule 6 "Triple Votes" to read as follows:
Every time a rule which is numbered with a multiple of 10 is added to the official-rules, all active players are given a TripleVote. Y6 N1

Proposal 40: Amend Rule 31: "Rule Titles" to the following:

Each rule may have a title in addition to its number. If the proposal creating a rule specifies a title for the new rule, then that becomes the rule's title upon the proposal's adoption. A proposal to amend a rule may specify a new title for the rule; that title becomes the rule's title of the proposal takes effect. If a rule ever lacks a title, any Github collaborator may give it a title.

If a proposal to create, amend or repeal a rule fails to include the title of the rule, 20 coins are transferred from the proposal creator to the Bank upon the proposal's adoption. If a proposal specifies a new title for a rule, the old title must be included as well to avoid the fine.

For the sake of clarity, the constitution's title is "The Constitution". All proposals that affect The Constitution must include the following note at the top:
(This is a constitutional proposal, and as such goes directly to chamber voting tomorrow. This proposal requires at least 75% for votes to pass)

If a proposal affecting the constitution does not include this note, the proposals creator will face the penalty as if they had failed to include the title of the rule. Y 4 N 1