Proposal 37 and  41
Proposal 41:
New feudal action:

"Gamble"
Message a number from 1 to 20 (in the channel #general), inclusively, then use the command !gamble. If the number replied by a bot is higher than the number you gambled, you gain as many Coins as you had gambled. 
If the number is not higher than the number you gambled, you lose as many Coins as you had gambled.
Note: the !gamble command would reply with a random number from 1 to 20, and also ping a mod. Y3 N1


Proposal 37: create a new feudal action:
Loaning: if a player takes this action, they may request to loan a specific number of coins from the bank. The amount of the loan must be less than or equal to the current amount in the bank and can be no higher than 200 coins.
Within the next 24 hours, the Merchant can choose to approve or deny the loan.
If it is denied, nothing more happens.
If it is approved, the amount of coins requested is removed from the bank and added to the coins of the player who requested the loan.
The player gains a loan value in the inventory equal to the amount of coins they got from the bank, is considered a "debtor" and gets the according role for as long as they have a non-zero loan value.

The merchant may not take this action.

The regulations about loans stated in rule 26 (loans) also apply to loans taken out as a feudal action. If one of those regulations conflict with the wording of this action, then the wording of this action takes precedence for loans taken out as a feudal action. Y4 N2