Proposal 44 y4n1
Proposal 44: Revamped Counties

Change the shop item "County in the Kingdom of Nomica" to allow non-Vassal players to purchase them for 150 coins. (Vassal players still can buy them for 100 coins.) In addition, every Friday at announcement time, increase the number of Counties in supply by 1.

The description of the County item is thus changed to:

Every day, a moderator generates a random number between 1 and 10. That is the income yielded by each county that day. Every owner of a county immediately gains that number of coins for each county he or she owns. 
A player who is not a vassal OR has vassals of their own must pay an additional 50 coins to the Bank to purchase this item. Every Friday at announcement time, the supply of this item is increased by 1.