proposal 43 / 46
Proposal 43: Weapons: At the time of this proposal becoming a rule every active player and future active players gets a weapon for uses to be specified by future rules. 

The weapon given by this rule will be a dagger. y4n1

Proposal 46: Change rule 10 to read as follows:

Instead of making a proposal, a @Current player may start a debate. To initiate a debate, a player submits the opening proposition that will be debated in the channel #current_debates. The same player should then immediately submit the opening statement in favour of the proposition in #debate_discussion titled "Opening Statement For the Proposition". The next input into the debate should be written by another player, a player who rejects the proposition. This input should be titled "Opening Statement Against the Proposition". The two sides then alternate rebutting each other's arguments, titling each input into the debate "Rebuttal X" (X being the next consecutive number for each side).

Debating continues on the day the debate is initiated and for two days after that. On the third day after the debate was initiated, each side submits an input titled "Closing Statement", summing up one's position and arguments. At the end of that last day, all players who submitted at least one rebuttal or statement receive 80 points. Furthermore, players can tell a moderator that a particular argument changed their mind on the matter at hand, for which the player that wrote the argument gains an additional 30 points. This should be done within a reasonable time after the debate, but not necessarily immediately after.

A player can report an input into the debate for being too low-effort. If a moderator considers the input to be trolling, not comprehensively written or not containing sound argument(s), the moderator shall remove rebuttal or statement in question and notify the author that so has been done. If not, the report is dropped and the concerned players are notified Y4 N2