Proposal 23 3-0
Proposal 23: Amending To Explore Strange New Worlds

Amend rule 11 to read:

A player may move their Ship to a coordinate pair which does not exist in the map if the coordinate pair consists of two integers AND the coordinate pair would be legal to move to if it did exist. When this happens, the player who moved to the new coordinate pair must Encounter a Card.