Proposals 26, 29-33
Proposal 26:  txt message

Replace
A player submits a proposal by posting a document labeled as such in #proposals.


in rule 2 with
A player submits a proposal by posting a text message labeled as such (which does not refer to a document) in #proposals.

Resu (13, 94)06/10/2020
Proposal 29: let's get that bread

Append the following to the end of the rule "Factories and Modules":
**Name:** The Moneymaker
**Cost:** 3 Dad Coins
**Description:** *You gotta spend money to make money.*
**Effect:** When equipping this module to a factory, pay 1 Dad Coin. When collecting from this module's factory, gain 3 Dad Coins.
Beowulf (11, §3)06/10/2020
Proposal 30: Old Pa's got your frack

Append the following to the end of the rule "Factories and Modules":
**Name:** Father Frack
**Cost:** 1 Dad Coin
**Description:** *Your trusty pa is back to help you frack that sweet sweet Dad Coin, make your pappy proud.*
**Effect:** When equipping this module to a factory, pay 2 points. When collecting from this module's factory, gain 2 Dad Coins.


Edits: Typo, trust to trusty
Resu (13, 94)06/10/2020
Proposal 31: Now Hiring

Amend the rule "Factories and Modules" by inserting the following paragraphs before the list of modules:
A player may hire other players to work in their factory. To do so, the player doing the hiring and the player being hired must both indicate their agreement factorially. If and when a player collects from their factory, they receive one additional point for each player in their employ, and each player in their employ receives one Dad Coin. 

A player may hire multiple other players to work in their factory, but no player may work in more than one factory at a time. A player may not hire themselves to work in their own factory. A player may fire any of their employees or quit their job by announcing it factorially.
The Cat (11, |3)06/10/2020
Proposal 32: Legislative Combo merge

Repeal the Rule with the title "Legistlative Combo".

Amend Rule 3 by replacing the sentence beginning "On a proposing day" with the following: "A player's proposal quota is 2 if, during their group's previous proposal day, they submitted a proposal that was later adopted, and otherwise 1.  On their group's proposing day, each member of that group may submit a number of proposals not exceeding their proposal quota; any excess proposal is illegal."
Proposal 33: Rule number clarification

If Rule 1 does not end in a period, amend it by appending a period to it.

Amend Rule 1 by appending the following: "No two rules can have the same ID number. If there is general consensus that a rule has a specific ID number, that rule's ID number is set to the consensus, regardless of what it previously was."

[This ensures that if we mess up rule ID numbers in tracking then nothing actually breaks.]